% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_search.R
\name{wiki_search}
\alias{wiki_search}
\title{Display results of a Wikipedia search in the browser}
\usage{
wiki_search(term = NULL, browser = getOption("browser"))
}
\arguments{
\item{term}{A non-empty character string giving the name of the term to be searched}

\item{browser}{A non-empty character string passed to [browseURL()] to determine the browser used.}
}
\value{
A display of the results of the search in the browser.
}
\description{
`wiki_search()` displays the results of a Wikipedia search in the browser.
}
\examples{
wiki_search('R (programming language)')
}
