% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_didyouknow.R
\name{wiki_didyouknow}
\alias{wiki_didyouknow}
\title{Generate random 'did you know' fact}
\usage{
wiki_didyouknow()
}
\value{
A message with a random 'did you know' fact from Wikipedia's main page
}
\description{
Generate random 'did you know' fact
}
\examples{
wiki_didyouknow()
}
