% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_inthenews.R
\name{wiki_inthenews}
\alias{wiki_inthenews}
\title{Generate random news item}
\usage{
wiki_inthenews()
}
\value{
A message with a random 'in the news' item from Wikipedia's main page
}
\description{
Generate random news item
}
\examples{
wiki_inthenews()
}
