% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_onthisday.R
\name{wiki_onthisday}
\alias{wiki_onthisday}
\title{Generate random 'on this day' fact}
\usage{
wiki_onthisday()
}
\value{
A message with a random 'on this day' fact from Wikipedia's main page
}
\description{
Generate random 'on this day' fact
}
\examples{
wiki_onthisday()
}
