% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_query.R
\name{wiki_query}
\alias{wiki_query}
\title{Send queries to the Wikidata knowledge graph and receive results as dataframe.}
\usage{
wiki_query(qry, raw_results = FALSE)
}
\arguments{
\item{qry}{A character string representing a SPARQL query to be run against the Wikidata knowledge graph.}

\item{raw_results}{Logical.  If TRUE, the results of the query will not be cleaned to remove escaped quotes and language tags.}
}
\value{
A dataframe of results
}
\description{
`wiki_query()` sends a SPARQL query to the Wikidata knowledge graph and collects the results in a dataframe.
}
\examples{
# List five diseases
library(SPARQL)
query <- 'SELECT ?item ?itemLabel WHERE {
  ?item wdt:P31 wd:Q12136. #instance of disease
  SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
}
LIMIT 5'
wiki_query(query)
}
