% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_randomfact.R
\name{wiki_randomfact}
\alias{wiki_randomfact}
\title{Generate random fact}
\usage{
wiki_randomfact()
}
\value{
A message with a random item from Wikipedia's main page
}
\description{
Generate random fact
}
